+++ core/src/definite/items/self_item.rs
@@ -3,6 +3,6 @@ use serde::Serialize;
 
 #[derive(PartialEq, Debug, Clone, Serialize, Deserialize)]
 pub struct SelfItem {
-    pub class_page: u64,
-    pub class_hash: u64,
+    pub class_page: usize,
+    pub class_hash: usize,
 }
