+++ core/src/definite/items/type_def.rs
@@ -4,5 +4,5 @@ use serde::{Deserialize, Serialize};
 #[derive(PartialEq, Debug, Clone, Serialize, Deserialize, Default)]
 pub struct TypeDef {
     pub name: String,
-    pub hash: u64,
+    pub hash: usize,
 }
